WIP bug 1365987: Rename regen to rerender
